move FullStatus tests to new file for consistency
